airflow/requirements/base.txt
astronomer-cosmos
apache-airflow-providers-docker